Add files via upload
Congratulations, we have made it to the end of this project! In this project, we covered a lot of details about Customer Segmentation. We have learned what the customer segmentation is, Need of Customer Segmentation, RFM analysis, Implementation of RFM from scratch in python. Also, we covered some basic concepts of pandas such as handling duplicates, groupby , and qcut () for bins based on sample quantiles. Thanks for reading this project tutorial.
 For any queries reach me out at amenasaiyed2516@gmail.com 
Amena Saiyed